# src/openapi_server/models/request_signup_model.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model_ import Model
from openapi_server import util


class RequestSignupModel(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, mail_address=None, password=None):  # noqa: E501
        """RequestSignupModel - a model defined in OpenAPI

        :param mail_address: The mail_address of this RequestSignupModel.  # noqa: E501
        :type mail_address: str
        :param password: The password of this RequestSignupModel.  # noqa: E501
        :type password: str
        """
        self.openapi_types = {
            'mail_address': str,
            'password': str
        }

        self.attribute_map = {
            'mail_address': 'mail_address',
            'password': 'password'
        }

        self._mail_address = mail_address
        self._password = password

    @classmethod
    def from_dict(cls, dikt) -> 'RequestSignupModel':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The RequestSignupModel of this RequestSignupModel.  # noqa: E501
        :rtype: RequestSignupModel
        """
        return util.deserialize_model(dikt, cls)

    @property
    def mail_address(self):
        """Gets the mail_address of this RequestSignupModel.


        :return: The mail_address of this RequestSignupModel.
        :rtype: str
        """
        return self._mail_address

    @mail_address.setter
    def mail_address(self, mail_address):
        """Sets the mail_address of this RequestSignupModel.


        :param mail_address: The mail_address of this RequestSignupModel.
        :type mail_address: str
        """

        self._mail_address = mail_address

    @property
    def password(self):
        """Gets the password of this RequestSignupModel.


        :return: The password of this RequestSignupModel.
        :rtype: str
        """
        return self._password

    @password.setter
    def password(self, password):
        """Sets the password of this RequestSignupModel.


        :param password: The password of this RequestSignupModel.
        :type password: str
        """

        self._password = password