# src/openapi_server/models/response_gen_key_model.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model_ import Model
from openapi_server import util


class ResponseGenKeyModel(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, public_key_hex=None, secret_key_hex=None):  # noqa: E501
        """ResponseGenKeyModel - a model defined in OpenAPI

        :param public_key_hex: The public_key_hex of this ResponseGenKeyModel.  # noqa: E501
        :type public_key_hex: str
        :param secret_key_hex: The secret_key_hex of this ResponseGenKeyModel.  # noqa: E501
        :type secret_key_hex: str
        """
        self.openapi_types = {
            'public_key_hex': str,
            'secret_key_hex': str
        }

        self.attribute_map = {
            'public_key_hex': 'public_key_hex',
            'secret_key_hex': 'secret_key_hex'
        }

        self._public_key_hex = public_key_hex
        self._secret_key_hex = secret_key_hex

    @classmethod
    def from_dict(cls, dikt) -> 'ResponseGenKeyModel':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The ResponseGenKeyModel of this ResponseGenKeyModel.  # noqa: E501
        :rtype: ResponseGenKeyModel
        """
        return util.deserialize_model(dikt, cls)

    @property
    def public_key_hex(self):
        """Gets the public_key_hex of this ResponseGenKeyModel.


        :return: The public_key_hex of this ResponseGenKeyModel.
        :rtype: str
        """
        return self._public_key_hex

    @public_key_hex.setter
    def public_key_hex(self, public_key_hex):
        """Sets the public_key_hex of this ResponseGenKeyModel.


        :param public_key_hex: The public_key_hex of this ResponseGenKeyModel.
        :type public_key_hex: str
        """

        self._public_key_hex = public_key_hex

    @property
    def secret_key_hex(self):
        """Gets the secret_key_hex of this ResponseGenKeyModel.


        :return: The secret_key_hex of this ResponseGenKeyModel.
        :rtype: str
        """
        return self._secret_key_hex

    @secret_key_hex.setter
    def secret_key_hex(self, secret_key_hex):
        """Sets the secret_key_hex of this ResponseGenKeyModel.


        :param secret_key_hex: The secret_key_hex of this ResponseGenKeyModel.
        :type secret_key_hex: str
        """

        self._secret_key_hex = secret_key_hex